feat: add watchlist (add/remove/list) and migrate monetary fields to BigDecimal

+++ src/main/java/com/tech/snapbid/dto/AuctionItemResponseDto.java
@@ -1,5 +1,6 @@
 package com.tech.snapbid.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import lombok.Data;
@@ -10,13 +11,13 @@ public class AuctionItemResponseDto {
     private Long id;
     private String title;
     private String description;
-    private Double startingPrice;
+    private BigDecimal startingPrice;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
     private String sellerUsername;
     private String status; 
     private String winnerUsername;
-    private Double finalPrice;
+    private BigDecimal finalPrice;
     private Long timeRemainingSeconds;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
