+++ src/main/java/com/tech/snapbid/dto/AuctionItemUpdateDto.java
@@ -1,5 +1,6 @@
 package com.tech.snapbid.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import lombok.Data;
@@ -9,7 +10,7 @@ public class AuctionItemUpdateDto {
 
     private String title;
     private String description;
-    private Double startingPrice;
+    private BigDecimal startingPrice;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
